feat(runtime): make target firmware fully freestanding and eliminate newlib/libnosys (#847)
Make target firmware fully freestanding and eliminate all implicit libc/newlib
and libnosys runtime dependencies in favor of audited project-owned primitives,
explicit compiler runtime support from libgcc, and strict post-link gates.

Key changes:
- Toolchain & Linker: migrate target link flags from --specs=nano.specs and
  --specs=nosys.specs to -nostdlib -Wl,--gc-sections with explicit trailing
  libgcc linkage for __aeabi_* compiler helpers. Eliminates libg_nano.a and
  libnosys.a completely from target links.
- Freestanding Runtime: implement project-owned, bounded, heap-free C runtime
  primitives in libs/ra8_core/ (memset, memcpy, memmove, memcmp, memchr,
  strlen, strnlen, strcmp, strncmp, strchr, strrchr, strstr, strcpy, strncpy,
  abs, labs) with GCC loop-distribution patterns disabled.
- Zero-Heap Invariant: verify no target linker script defines 'end' or '_end'
  (removed from compile_on_m33/linker_script_cpu1.ld) or '.heap'. Accidental
  calls to malloc/calloc/realloc/free fail closed at link time.
- Correctness & Hardening: protect calloc multiplication against integer
  overflow in tls_client and threadx_https_client callbacks. Replace runtime
  libc assert() in ra8_jpeg_sw_decode and ra8_rabook_gray4 with RA8_ASSERT.
  Remove dead <assert.h> includes from 5 headers/TUs.
- Vendored SOUP: activate MINIZ_NO_MALLOC, MINIZ_NO_STDIO, MINIZ_NO_TIME on
  target miniz and guard MZ_ASSERT to expand to ((void)0), cutting the stdio
  and assert dependency chain.
- CI & Gating: add scripts/checks/check_freestanding_runtime.py post-link binary
  and map checker with --check-asserts and --check-scripts, wired into
  pre-commit-checks and just apps::build. Added unit test suites for calloc
  guards and freestanding runtime primitives.

+++ libs/ra8_core/inc/ra8_sbrk_trap.h
@@ -7,18 +7,11 @@
  * [Ring 1 / CORE] {World: S}
  *
  * @details
- * NASA Power of 10 Rule 3 forbids dynamic allocation after init, and this
- * firmware has no heap at all. ``ra8_sbrk_trap.c`` defines a strong
- * ``_sbrk`` so any code that reaches newlib's allocator halts loudly at
- * the exact call rather than silently corrupting whatever sits past the
- * break.
- *
- * The symbol is external because the C library resolves it -- newlib's
- * ``malloc`` calls ``_sbrk`` by name, and nothing in this tree calls it
- * at all. This header exists so the definition and the host death-test
- * that drives it agree on one prototype instead of restating it: they had
- * a copy each, and nothing checked them against each other.
- *
+ * NASA Power of 10 Rule 3 forbids dynamic allocation after init, and target
+ * firmware has no heap at all. Firmware is freestanding and links with `-nostdlib`
+ * without newlib or libnosys. Standard allocator calls fail closed at link time.
+ * ``ra8_sbrk_trap.c`` provides a defense-in-depth tripwire so any legacy or
+ * external routine attempting to invoke ``_sbrk`` halts safely via ``ra8_fatal_error``.
  * @copyright Copyright (c) 2026 Brighton Sikarskie
  * SPDX-License-Identifier: MIT
  */
@@ -29,7 +22,7 @@
 extern "C" {
 #endif
 
-#include <stdint.h>
+#include <stddef.h>
 
 // NOLINTBEGIN(bugprone-reserved-identifier,cert-dcl37-c,cert-dcl51-cpp,readability-identifier-naming) -- toolchain ABI fixes this reserved symbol spelling.
 
@@ -69,7 +62,7 @@ extern "C" {
  *
  * @since 0.1.0
  */
-void* _sbrk(int32_t incr);
+void* _sbrk(ptrdiff_t incr);
 
 // NOLINTEND(bugprone-reserved-identifier,cert-dcl37-c,cert-dcl51-cpp,readability-identifier-naming)
